Create fixissues.ps1
This was the ps1 file I was using to inject intrin_workarounds.h into files.

I am pushing it here to have it preserved in case we need it in the future, or more likely, to use as a basis in the future in order to easily remove the injection.